requirements.txt
certifi==2020.6.20
psycopg2==2.8.6
pyaml==20.4.0
PyYAML==5.3.1
SQLAlchemy==1.3.20
